refactor: extracted reducer into a new file
